Updated commands to fix lost functionality
adding commands means we loose our root comand, this got lost during recent development.
